// models/FoundDictionaryItemDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kupsilla Data Platform dictionary engine as a service
 * Swagger for \'Kupsilla Data Platform dictionary engine as a service\'
 *
 * The version of the OpenAPI document: 0.1.0-df9ccc8-df9ccc8-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    TextContentDTO,
} from './TextContentDTO.js';

/**
 * Response with information about the dictionary item
 * @export
 * @interface FoundDictionaryItemDTO
 */
export interface FoundDictionaryItemDTO {
    /**
     * Unique comprehensible dictionary type
     * @type {string}
     * @memberof FoundDictionaryItemDTO
     */
    dictionary_type: string;
    /**
     * The path to parent dictionary item, represented as a sequence of item keys that begins and is separated by a slash (/)
     * @type {string}
     * @memberof FoundDictionaryItemDTO
     */
    parent_item_path: string;
    /**
     * A comprehensible key of a dictionary element that is unique among elements of the same node
     * @type {string}
     * @memberof FoundDictionaryItemDTO
     */
    item_key: string;
    /**
     * The level of this dictionary item in the dictionary tree
     * @type {number}
     * @memberof FoundDictionaryItemDTO
     */
    level: number;
    /**
     * Whether this dictionary item can be selected or not
     * @type {boolean}
     * @memberof FoundDictionaryItemDTO
     */
    selectable: boolean;
    /**
     * 
     * @type {TextContentDTO}
     * @memberof FoundDictionaryItemDTO
     */
    item_content: TextContentDTO;
    /**
     * Collection of child items (collection of SearchDictionaryItemResponseDTO objects)
     * @type {Array<FoundDictionaryItemDTO>}
     * @memberof FoundDictionaryItemDTO
     */
    items?: Array<FoundDictionaryItemDTO>;
}